Roomies/requirements.txt
asgiref==3.8.1
distlib==0.3.9
Django==5.2
django-cors-headers==4.7.0
django-localflavor==4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
filelock==3.18.0
mysqlclient==2.2.7
pillow==11.2.1
platformdirs==4.3.7
PyJWT==2.9.0
python-stdnum==1.20
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.30.0
